refactor(observability): 归位 Skill Health 能力
