+++ test/managed/observe-feedback.test.ts
@@ -11,8 +11,8 @@ import { tmpdir } from 'node:os';
 import { buildManagedArtifactRecord, upsertManagedRecord, loadManagedRecord, managedRecordId } from '../../src/managed/store.js';
 import { recordObserveHealth, type ObservedSkillHealthView } from '../../src/managed/observe-feedback.js';
 import { buildObserveReportView } from '../../src/cli/commands/observe/index.js';
-import { healthBandOf } from '../../src/observability/skill-health-analyzer.js';
-import type { SkillHealthReport } from '../../src/observability/skill-health-analyzer.js';
+import { healthBandOf } from '../../src/observability/skill-health/analyzer.js';
+import type { SkillHealthReport } from '../../src/observability/skill-health/analyzer.js';
 import type { ArtifactKind } from '../../src/artifacts/contracts.js';
 
 const GAP0 = { failed_search: 0, explicit_marker: 0, hedging: 0, repeated_failure: 0 };
