+++ test/observability/soft-standards.test.ts
@@ -6,7 +6,7 @@ import {
   type RuntimeSignalType,
   type RuntimeStandardNode,
 } from '../../src/observability/soft-standards/index.js';
-import type { SkillRuntimeEvidencePack, SkillRuntimeEvidencePackRef, SkillRuntimeEvidencePackSourceType } from '../../src/observability/skill-chain.js';
+import type { SkillRuntimeEvidencePack, SkillRuntimeEvidencePackRef, SkillRuntimeEvidencePackSourceType } from '../../src/observability/skill-health/skill-chain.js';
 
 const { evaluateRuntimeStandardNodes, normalizeRuntimeSignals, normalizeRuntimeTriggers, normalizeFuzzyText, normalizeToolNameValue } = __softStandardsTestInternals;
